+++ src/urls.rs
@@ -332,11 +332,11 @@ pub struct LongestMatchingPrefixValues<'a, V> {
     current_node: Option<usize>,
 }
 
-impl<V> LongestMatchingPrefixValues<'_, V> {
-    pub fn has_next(&self) -> bool {
-        self.current_node.is_some()
-    }
-}
+// impl<V> LongestMatchingPrefixValues<'_, V> {
+//     pub fn has_next(&self) -> bool {
+//         self.current_node.is_some()
+//     }
+// }
 
 impl<'a, V> Iterator for LongestMatchingPrefixValues<'a, V> {
     type Item = &'a V;